A Simple Namespace
I modify the structure to avoid any namespace clashes with preinstalled
software, all the structure on the Autorig change, was a huge work.
Changes on Name convention, now the name convention is stronger to be
able to change from project to project. the RMNameConvention had some
changes that affected all the tools, so everithing had to be modified,
this is a stable version where everithing should be fixed.