Several fixes
Fixed a problem with the Reverse spine, that it was rotating when the back was rotated.
Fixed a problem with the scaling in general. Now the rig should be scalable. (it requires one small change on the biped build.
Fixed a problem when loading shapes with multiple shapes. For example the Bs, changing this to a circle and loading it was not doing it correctly.